backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
stripe==10.5.0
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9; python_version < "3.14"
Pillow>=11.0.0
python-dotenv==1.0.1
